requirements.txt
## Core web

fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
websockets>=12.0
httptools>=0.6.0
uvloop>=0.19.0; sys_platform != "win32"
watchfiles>=0.21.0

## Async HTTP (prefer for async backends) and sync fallback

httpx>=0.26.0
requests>=2.31.0
Brotli>=1.1.0

## Environment

python-dotenv>=1.0.0

## Database & ORM (async support)
SQLAlchemy>=2.0.25
asyncpg>=0.29.0    # async Postgres driver
psycopg2-binary>=2.9.9  # for sync tooling / migrations
alembic>=1.13.1

## File uploads / static handling

aiofiles>=23.2.1
python-multipart>=0.0.7

## Validation and models

pydantic>=2.5.3,<3.0.0
pydantic-settings>=2.1.0

## Authentication / security

passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

## Logging & monitoring

loguru>=0.7.2
sentry-sdk>=1.40.0

## LLM / Agent integrations

google-generativeai>=0.3.2
langchain>=0.1.0         # common LLM orchestration library
langchain-core>=0.1.10
langchain-google-genai>=0.0.6
langgraph>=0.0.10

## Production process manager (use with caution; choose one that fits your infra)

gunicorn>=21.2.0

## Dev / Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

## httpx is already listed above, useful for tests too

## Utilities

python-dateutil>=2.8.2
typing-extensions>=4.9.0

## Optional: add any other agent-specific connectors or SDKs here

## e.g. redis, celery, rq, etc.
